Moved some enums into utilities
