+++ Assets/MixedRealityToolkit/_Core/Definitions/Devices/InteractionDefinition.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using Microsoft.MixedReality.Toolkit.Internal.Definitions.InputSystem;
+using Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities;
 using System;
 using UnityEngine;
