+++ Assets/MixedRealityToolkit/_Core/Definitions/Utilities/AxisType.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace Microsoft.MixedReality.Toolkit.Internal.Definitions
+namespace Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities
 {
     /// <summary>
     /// The AxisType identifies the type of button or input being sent to the framework from a controller.
