+++ Assets/MixedRealityToolkit/_Core/Utilities/MotionController/MotionControllerVisualizer.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
-using Microsoft.MixedReality.Toolkit.Internal.Definitions;
+using Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities;
 using UnityEngine;
 
 #if UNITY_EDITOR_WIN
